requirements.txt
# Core dependencies
streamlit>=1.32.0
loguru>=0.7.2
pydantic>=2.6.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.3

# LLM provider dependencies
litellm>=1.30.7

# Data processing and analysis
pandas>=2.2.0
numpy>=1.26.3
scikit-learn>=1.4.0
plotly>=5.18.0

# Web scraping and document processing
beautifulsoup4>=4.12.3
selenium>=4.18.1
webdriver-manager>=4.0.1
python-docx>=1.1.0
PyPDF2>=3.0.1

# Caching and performance
diskcache>=5.6.3
lru-dict>=1.2.0

# Testing and development
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0 